+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 - `no_overlap` argument in `sopa.aggregate` to avoid cells from overlapping when aggregating the channels
-- Add `unique_mapping` argument to map each bin to one unique cell for Visium HD data
+- Add `no_overlap` argument to map each bin to one unique cell for Visium HD data using PCA proximity
 - Better documentation for `sopa.io.visium_hd` and a warning if the full res image is not loaded (#254)
 - Support `CONCH` for H&E patches inference.
 - Support `cellpose>=4.0.0` @lguerard (#252, #264)